Product And CategoryDetail

+++ src/main/java/com/example/demo/dao/AccountDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.AccountEntity;
 
 
-public interface AccountDAO extends JpaRepository<AccountEntity, Long>{
+public interface AccountDAO extends JpaRepository<AccountEntity, Integer>{
 
 }
